rename register
