+++ controllers/users/index.js
@@ -1,11 +1,11 @@
-const signup = require("./signup");
+const register = require("./register");
 const login = require("./login");
 const logout = require("./logout");
 const getCurrent = require("./getCurrent");
 const updateSubscription = require("./updateSubscription");
 
 module.exports = {
-  signup,
+  register,
   login,
   logout,
   getCurrent,
